feat(GenericContainer): add StartupSetting trait for managing container startup timeout and check strategy
