+++ tests/Unit/Containers/GenericContainerTest.php
@@ -3,11 +3,8 @@
 namespace Tests\Unit\Containers;
 
 use PHPUnit\Framework\TestCase;
-use Symfony\Component\Process\Exception\ProcessTimedOutException;
-use Testcontainers\Containers\BindMode;
 use Testcontainers\Containers\ContainerInstance;
 use Testcontainers\Containers\GenericContainer\GenericContainer;
-use Testcontainers\Containers\ImagePullPolicy;
 use Testcontainers\Containers\PortStrategy\LocalRandomPortStrategy;
 use Testcontainers\Containers\WaitStrategy\LogMessageWaitStrategy;
 use Testcontainers\Docker\DockerClientFactory;
@@ -99,16 +96,5 @@ class GenericContainerTest extends TestCase
         $this->assertTrue(is_int($instance->getMappedPort(443)));
         $this->assertGreaterThanOrEqual(49152, $instance->getMappedPort(443));
         $this->assertLessThanOrEqual(65535, $instance->getMappedPort(443));
-    }
-
-    public function testStartWithStartupTimeout()
-    {
-        $this->expectException(ProcessTimedOutException::class);
-
-        $container = (new GenericContainer('alpine:latest'))
-            ->withStartupTimeout(1)
-            ->withCommands(['sleep', '5']);
-        /** @noinspection PhpUnhandledExceptionInspection */
-        $container->start();
     }
 }
